# QL_FOODDELI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QL_FOODDELI API (Dapper & SignalR)",
    "description": "Hệ thống backend API quản lý giao đồ ăn nhanh QL_FOODDELI - Sprint 5",
    "version": "v1"
  },
  "paths": {
...
    "/api/DanhMuc": {
      "get": {
        "tags": [
          "DanhMuc"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "DanhMuc"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DanhMuc"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DanhMuc"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DanhMuc"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DanhMuc/{id}": {
      "get": {
        "tags": [
          "DanhMuc"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "DanhMuc"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DanhMuc"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DanhMuc"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DanhMuc"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "DanhMuc"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/DiaChi": {
      "post": {
        "tags": [
          "DiaChi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaChiCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaChiCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DiaChiCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "DiaChi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaChiUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaChiUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DiaChiUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DiaChi/{maDiaChi}/mac-dinh": {
      "put": {
        "tags": [
          "DiaChi"
        ],
        "parameters": [
          {
            "name": "maDiaChi",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DiaChi/{maDiaChi}": {
      "delete": {
        "tags": [
          "DiaChi"
        ],
        "parameters": [
          {
            "name": "maDiaChi",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DanhMuc": {
        "type": "object",
        "properties": {
          "maDanhMuc": {
            "type": "string",
            "nullable": true
          },
          "tenDanhMuc": {
            "type": "string",
            "nullable": true
          },
          "moTa": {
            "type": "string",
            "nullable": true
          },
          "trangThai": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DiaChiCreateRequest": {
        "type": "object",
        "properties": {
          "maDiaChi": {
            "type": "string",
            "nullable": true
          },
          "nguoiNhan": {
            "type": "string",
            "nullable": true
          },
          "soDienThoai": {
            "type": "string",
            "nullable": true
          },
          "diaChiChiTiet": {
            "type": "string",
            "nullable": true
          },
          "phuongXa": {
            "type": "string",
            "nullable": true
          },
          "quanHuyen": {
            "type": "string",
            "nullable": true
          },
          "tinhThanh": {
            "type": "string",
            "nullable": true
          },
          "macDinh": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DiaChiUpdateRequest": {
        "type": "object",
        "properties": {
          "nguoiNhan": {
            "type": "string",
            "nullable": true
          },
          "soDienThoai": {
            "type": "string",
            "nullable": true
          },
          "diaChiChiTiet": {
            "type": "string",
            "nullable": true
          },
          "phuongXa": {
            "type": "string",
            "nullable": true
          },
          "quanHuyen": {
            "type": "string",
            "nullable": true
          },
          "tinhThanh": {
            "type": "string",
            "nullable": true
          },
          "macDinh": {
            "type": "boolean"
          },
          "maDiaChi": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Dán mã JWT Token thu được sau khi Đăng nhập vào ô này (KHÔNG gõ chữ Bearer):",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}